+++ src/Attachment.php
@@ -4,7 +4,7 @@
  *
  * @license http://opensource.org/licenses/MIT
  * @link https://github.com/carpediem/mattermost-php/
- * @version 1.0.1
+ * @version 1.1.0
  * @package carpediem.mattermost-webhook
  *
  * For the full copyright and license information, please view the LICENSE
@@ -127,6 +127,37 @@ final class Attachment implements JsonSerializable
      */
     private $thumb_url = '';
 
+    /**
+     * Returns a new instance from an array.
+     *
+     * @param array $arr
+     *
+     * @return self
+     */
+    public static function fromArray(array $arr)
+    {
+        $prop = $arr + (new self())->toArray();
+
+        return self::__set_state($prop);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public static function __set_state(array $prop)
+    {
+        return (new self())
+            ->setFallback($prop['fallback'])
+            ->setColor($prop['color'])
+            ->setPretext($prop['pretext'])
+            ->setText($prop['text'])
+            ->setTitle($prop['title'], $prop['title_link'])
+            ->setAuthor($prop['author_name'], $prop['author_link'], $prop['author_icon'])
+            ->setFields($prop['fields'])
+            ->setImageUrl($prop['image_url'])
+            ->setThumbUrl($prop['thumb_url'])
+        ;
+    }
 
     /**
      * {@inheritdoc}
@@ -227,18 +258,31 @@ final class Attachment implements JsonSerializable
     }
 
     /**
-     * @param string $author_name
+     * Set attachment author.
      *
-     * @return self
+     * @param string              $author_name
+     * @param string|UriInterface $author_link
+     * @param string|UriInterface $author_icon
      */
-    public function setAuthorName($author_name)
+    public function setAuthor($author_name, $author_link = '', $author_icon = '')
     {
         $this->author_name = filter_string($author_name, 'author_name');
+        if ('' === $this->author_name) {
+            $this->author_link = '';
+            $this->author_icon = '';
+
+            return $this;
+        }
+
+        $this->author_link = filter_uri($author_link, 'author_link');
+        $this->author_icon = filter_uri($author_icon, 'author_icon');
 
         return $this;
     }
 
     /**
+     * Returns the author name.
+     *
      * @return string
      */
     public function getAuthorName()
@@ -247,18 +291,8 @@ final class Attachment implements JsonSerializable
     }
 
     /**
-     * @param string|UriInterface $author_link
+     * Returns the author link URI.
      *
-     * @return self
-     */
-    public function setAuthorLink($author_link)
-    {
-        $this->author_link = filter_uri($author_link, 'author_link');
-
-        return $this;
-    }
-
-    /**
      * @return string
      */
     public function getAuthorLink()
@@ -267,23 +301,64 @@ final class Attachment implements JsonSerializable
     }
 
     /**
-     * @param string|UriInterface $author_icon
+     * Returns the author icon URI.
      *
-     * @return self
-     */
-    public function setAuthorIcon($author_icon)
-    {
-        $this->author_icon = filter_uri($author_icon, 'author_icon');
-
-        return $this;
-    }
-
-    /**
      * @return string
      */
     public function getAuthorIcon()
     {
         return $this->author_icon;
+    }
+
+    /**
+     * Sets the author name.
+     *
+     * @param string $author_name
+     *
+     * DEPRECATION WARNING! This method will be removed in the next major point release
+     *
+     * @deprecated deprecated since version 1.1.0
+     * @see Attachment::setAuthor
+     *
+     * @return self
+     */
+    public function setAuthorName($author_name)
+    {
+        return $this->setAuthor($author_name, $this->author_link, $this->author_icon);
+    }
+
+    /**
+     * Sets the author link URI.
+     *
+     * @param string|UriInterface $author_link
+     *
+     * DEPRECATION WARNING! This method will be removed in the next major point release
+     *
+     * @deprecated deprecated since version 1.1.0
+     * @see Attachment::setAuthor
+     *
+     * @return self
+     */
+    public function setAuthorLink($author_link)
+    {
+        return $this->setAuthor($this->author_name, $author_link, $this->author_icon);
+    }
+
+    /**
+     * Sets the author name icon URI.
+     *
+     * @param string|UriInterface $author_icon
+     *
+     * DEPRECATION WARNING! This method will be removed in the next major point release
+     *
+     * @deprecated deprecated since version 1.1.0
+     * @see Attachment::setAuthor
+     *
+     * @return self
+     */
+    public function setAuthorIcon($author_icon)
+    {
+        return $this->setAuthor($this->author_name, $this->author_link, $author_icon);
     }
 
     /**